requirements.txt
flask
torch
torchvision
pillow
gunicorn

flask-compress
